refactor - 패키지구조 변경, RandomBooleanGenerator 로직 변경
