+++ src/test/java/nextstep/ladder/domain/LineTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.Line;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
